+++ src/models/Settings.php
@@ -27,6 +27,7 @@ class Settings extends Model
     public string $mjmlCliConfigArgs = '';
     public string $appId = '';
     public string $secretKey = '';
+    public string $apiUrl = '';
 
     public function defineBehaviors(): array
     {
@@ -39,6 +40,7 @@ class Settings extends Model
                     'mjmlCliConfigArgs',
                     'appId',
                     'secretKey',
+                    'apiUrl',
                 ],
             ],
         ];
@@ -47,7 +49,7 @@ class Settings extends Model
     public function defineRules(): array
     {
         return [
-            [['appId', 'secretKey', 'mjmlCliPath', 'nodePath', 'mjmlCliConfigArgs'], 'string'],
+            [['appId', 'secretKey', 'mjmlCliPath', 'nodePath', 'mjmlCliConfigArgs', 'apiUrl'], 'string'],
             [['appId', 'secretKey'], 'required', 'when' => fn(Settings $model) => !empty($model->appId) || !empty($model->secretKey)],
             [['mjmlCliPath', 'nodePath'], 'required', 'when' => fn(Settings $model) => !empty($model->mjmlCliPath) || !empty($model->nodePath)],
         ];
@@ -58,6 +60,7 @@ class Settings extends Model
         return [
             'appId' => Craft::t('mjml', 'API App ID'),
             'secretKey' => Craft::t('mjml', 'API Secret Key'),
+            'apiUrl' => Craft::t('mjml', 'API URL'),
             'mjmlCliPath' => Craft::t('mjml', 'MJML Cli Path'),
             'nodePath' => Craft::t('mjml', 'Node.js Path'),
             'mjmlCliConfigArgs' => Craft::t('mjml', 'MJML Cli Config Arguments'),